requirements.txt
numpy==1.13.3
opencv-python==3.3.0.10
pandas==0.23.1
Pillow==4.3.0
scikit-image==0.13.1
scipy==1.0.0
tensorflow-gpu==1.8.0
tqdm==4.19.8